# morphio-io/frontend/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Morphio.io API",
		"description": "This API will blow your mind",
		"version": "1.0.0"
	},
	"paths": {
...
		"/content/save-content": {
			"post": {
				"tags": ["Content", "Content"],
				"summary": "Save Content",
				"operationId": "save_content",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ContentCreate"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel_ContentOut_"
								}
							}
						}
					},
					"201": {
						"description": "Content saved successfully",
						"content": {
							"application/json": {
								"example": {
									"status": "success",
									"message": "Content saved successfully",
									"data": {
										"id": 1,
										"title": "My Awesome Content",
										"content": "This is some great content!",
										"created_at": "2023-04-01T12:00:00Z",
										"updated_at": "2023-04-01T12:00:00Z"
									}
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"example": {
									"detail": "Not authenticated"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"example": {
									"detail": "Rate limit exceeded. Try again in 60 seconds."
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"example": {
									"detail": "An unexpected error occurred."
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				]
			}
		},
		"/content/get-contents": {
			"get": {
				"tags": ["Content", "Content"],
				"summary": "Get Contents",
				"operationId": "list_contents",
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"parameters": [
					{
						"name": "page",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"minimum": 1,
							"default": 1,
							"title": "Page"
						}
					},
					{
						"name": "per_page",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"maximum": 100,
							"minimum": 1,
							"default": 5,
							"title": "Per Page"
						}
					},
					{
						"name": "template_id",
						"in": "query",
						"required": false,
						"schema": {
							"anyOf": [
								{
									"type": "integer"
								},
								{
									"type": "null"
								}
							],
							"description": "Filter by template ID",
							"title": "Template Id"
						},
						"description": "Filter by template ID"
					}
				],
				"responses": {
					"200": {
						"description": "Contents retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel_PaginatedResponse_ContentOut__"
								},
								"example": {
									"status": "success",
									"message": "Contents retrieved successfully",
									"data": {
										"items": [
											{
												"id": 1,
												"title": "Content 1",
												"content": "Body of content 1",
												"created_at": "2023-04-01T12:00:00Z",
												"updated_at": "2023-04-01T12:00:00Z"
											}
										],
										"total": 10,
										"page": 1,
										"per_page": 2
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"example": {
									"detail": "Not authenticated"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"example": {
									"detail": "Rate limit exceeded. Try again in 60 seconds."
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"example": {
									"detail": "An unexpected error occurred."
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/content/get-content/{content_id}": {
			"get": {
				"tags": ["Content", "Content"],
				"summary": "Get Content",
				"operationId": "get_content",
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"parameters": [
					{
						"name": "content_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"title": "Content Id"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Content retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ContentOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"example": {
									"detail": "Not authenticated"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"example": {
									"detail": "Rate limit exceeded. Try again in 60 seconds."
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"example": {
									"detail": "An unexpected error occurred."
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/content/update-content/{content_id}": {
			"put": {
				"tags": ["Content", "Content"],
				"summary": "Update Content",
				"operationId": "update_content",
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"parameters": [
					{
						"name": "content_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"title": "Content Id"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ContentUpdate"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Content updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ContentOut"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"example": {
									"detail": "Not authenticated"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"example": {
									"detail": "Rate limit exceeded. Try again in 60 seconds."
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"example": {
									"detail": "An unexpected error occurred."
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/content/update-multiple-contents": {
			"put": {
				"tags": ["Content", "Content"],
				"summary": "Update Multiple Contents",
				"operationId": "update_multiple_contents",
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/ContentUpdate"
								},
								"type": "array",
								"title": "Contents"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Contents updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ContentOut"
									},
									"type": "array",
									"title": "Response Update Multiple Contents"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"example": {
									"detail": "Not authenticated"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"example": {
									"detail": "Rate limit exceeded. Try again in 60 seconds."
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"example": {
									"detail": "An unexpected error occurred."
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				},
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				]
			}
		},
		"/content/update-title/{content_id}": {
			"put": {
				"tags": ["Content", "Content"],
				"summary": "Update Content Title Route",
				"operationId": "update_content_title",
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"parameters": [
					{
						"name": "content_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"title": "Content Id"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ContentTitleUpdate"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Content title updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseModel_ContentOut_"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"example": {
									"detail": "Not authenticated"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"example": {
									"detail": "Rate limit exceeded. Try again in 60 seconds."
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"example": {
									"detail": "An unexpected error occurred."
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/content/delete-content/{content_id}": {
			"delete": {
				"tags": ["Content", "Content"],
				"summary": "Delete Content",
				"operationId": "delete_content",
				"security": [
					{
						"OAuth2PasswordBearer": []
					}
				],
				"parameters": [
					{
						"name": "content_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "integer",
							"title": "Content Id"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Content deleted successfully",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"example": {
									"detail": "Not authenticated"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"example": {
									"detail": "Content not found or unauthorized"
								}
							}
						}
					},
					"429": {
						"description": "Too Many Requests",
						"content": {
							"application/json": {
								"example": {
									"detail": "Rate limit exceeded. Try again in 60 seconds."
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"example": {
									"detail": "An unexpected error occurred."
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"ContentCreate": {
				"properties": {
					"title": {
						"type": "string",
						"title": "Title"
					},
					"content": {
						"type": "string",
						"title": "Content"
					},
					"user_id": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "User Id"
					},
					"template_id": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Template Id"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"title": "Tags",
						"default": []
					}
				},
				"type": "object",
				"required": ["title", "content"],
				"title": "ContentCreate"
			},
			"ContentOut": {
				"properties": {
					"title": {
						"type": "string",
						"maxLength": 200,
						"minLength": 1,
						"title": "Title"
					},
					"content": {
						"type": "string",
						"minLength": 1,
						"title": "Content"
					},
					"is_published": {
						"type": "boolean",
						"title": "Is Published",
						"default": false
					},
					"id": {
						"type": "integer",
						"title": "Id"
					},
					"user_id": {
						"type": "integer",
						"title": "User Id"
					},
					"template_id": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Template Id"
					},
					"created_at": {
						"type": "string",
						"format": "date-time",
						"title": "Created At"
					},
					"updated_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "null"
							}
						],
						"title": "Updated At"
					},
					"view_count": {
						"type": "integer",
						"title": "View Count",
						"default": 0
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"title": "Tags"
					},
					"user": {
						"$ref": "#/components/schemas/app__schemas__auth_schema__UserOut"
					},
					"template": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/TemplateOut"
							},
							{
								"type": "null"
							}
						]
					}
				},
				"type": "object",
				"required": [
					"title",
					"content",
					"id",
					"user_id",
					"created_at",
					"tags",
					"user"
				],
				"title": "ContentOut"
			},
			"ContentTitleUpdate": {
				"properties": {
					"title": {
						"type": "string",
						"maxLength": 200,
						"minLength": 1,
						"title": "Title"
					}
				},
				"type": "object",
				"required": ["title"],
				"title": "ContentTitleUpdate"
			},
			"ContentUpdate": {
				"properties": {
					"id": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Id"
					},
					"title": {
						"anyOf": [
							{
								"type": "string",
								"maxLength": 200,
								"minLength": 1
							},
							{
								"type": "null"
							}
						],
						"title": "Title"
					},
					"content": {
						"anyOf": [
							{
								"type": "string",
								"minLength": 1
							},
							{
								"type": "null"
							}
						],
						"title": "Content"
					},
					"is_published": {
						"anyOf": [
							{
								"type": "boolean"
							},
							{
								"type": "null"
							}
						],
						"title": "Is Published"
					},
					"tags": {
						"anyOf": [
							{
								"items": {
									"type": "string"
								},
								"type": "array"
							},
							{
								"type": "null"
							}
						],
						"title": "Tags"
					}
				},
				"type": "object",
				"title": "ContentUpdate"
			},
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": {
							"$ref": "#/components/schemas/ValidationError"
						},
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
...
			"PaginatedResponse_ContentOut_": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/ContentOut"
						},
						"type": "array",
						"title": "Items"
					},
					"total": {
						"type": "integer",
						"title": "Total"
					},
					"page": {
						"type": "integer",
						"title": "Page"
					},
					"per_page": {
						"type": "integer",
						"title": "Per Page"
					}
				},
				"type": "object",
				"required": ["items", "total", "page", "per_page"],
				"title": "PaginatedResponse[ContentOut]"
			},
...
			"ResponseModel_ContentOut_": {
				"properties": {
					"status": {
						"type": "string",
						"title": "Status"
					},
					"message": {
						"type": "string",
						"title": "Message"
					},
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ContentOut"
							},
							{
								"type": "null"
							}
						]
					}
				},
				"type": "object",
				"required": ["status", "message"],
				"title": "ResponseModel[ContentOut]"
			},
...
			"ResponseModel_PaginatedResponse_ContentOut__": {
				"properties": {
					"status": {
						"type": "string",
						"title": "Status"
					},
					"message": {
						"type": "string",
						"title": "Message"
					},
					"data": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/PaginatedResponse_ContentOut_"
							},
							{
								"type": "null"
							}
						]
					}
				},
				"type": "object",
				"required": ["status", "message"],
				"title": "ResponseModel[PaginatedResponse[ContentOut]]"
			},
...
			"TemplateOut": {
				"properties": {
					"id": {
						"type": "integer",
						"title": "Id"
					},
					"name": {
						"type": "string",
						"title": "Name"
					},
					"template_content": {
						"type": "string",
						"title": "Template Content"
					},
					"user_id": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "User Id"
					},
					"is_default": {
						"type": "boolean",
						"title": "Is Default"
					},
					"created_at": {
						"type": "string",
						"format": "date-time",
						"title": "Created At"
					},
					"updated_at": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "null"
							}
						],
						"title": "Updated At"
					},
					"usage_count": {
						"type": "integer",
						"title": "Usage Count",
						"default": 0
					},
					"user": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/app__schemas__auth_schema__UserOut"
							},
							{
								"type": "null"
							}
						]
					}
				},
				"type": "object",
				"required": [
					"id",
					"name",
					"template_content",
					"is_default",
					"created_at"
				],
				"title": "TemplateOut"
			},
...
			"UserRole": {
				"type": "string",
				"enum": ["USER", "ADMIN", "POWER", "MODERATOR", "GUEST"],
				"title": "UserRole"
			},
...
			"ValidationError": {
				"properties": {
					"loc": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "integer"
								}
							]
						},
						"type": "array",
						"title": "Location"
					},
					"msg": {
						"type": "string",
						"title": "Message"
					},
					"type": {
						"type": "string",
						"title": "Error Type"
					}
				},
				"type": "object",
				"required": ["loc", "msg", "type"],
				"title": "ValidationError"
			},
			"app__schemas__auth_schema__UserOut": {
				"properties": {
					"email": {
						"type": "string",
						"format": "email",
						"title": "Email"
					},
					"display_name": {
						"type": "string",
						"maxLength": 50,
						"minLength": 1,
						"title": "Display Name"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole",
						"default": "USER"
					},
					"id": {
						"type": "integer",
						"title": "Id"
					},
					"created_at": {
						"type": "string",
						"format": "date-time",
						"title": "Created At"
					},
					"last_login": {
						"anyOf": [
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "null"
							}
						],
						"title": "Last Login"
					},
					"is_active": {
						"type": "boolean",
						"title": "Is Active"
					},
					"content_count": {
						"type": "integer",
						"title": "Content Count",
						"default": 0
					}
				},
				"type": "object",
				"required": ["email", "display_name", "id", "created_at", "is_active"],
				"title": "UserOut"
			}
		},
		"securitySchemes": {
			"OAuth2PasswordBearer": {
				"type": "oauth2",
				"flows": {
					"password": {
						"scopes": {},
						"tokenUrl": "/auth/login"
					}
				}
			}
		}
	}
}